fix memset mistake and make api better for insert

+++ include/malgos/intrusive/hashtable.h
@@ -31,7 +31,7 @@ struct mlg_hash_table_s
 };
 
 mlg_error_t mlg_hashtable_init(mlg_hash_table_t *hash_table, mlg_hash_head_t *buckets, size_t size);
-mlg_error_t mlg_hashtable_insert(mlg_hash_table_t *hash_table, mlg_hash_node_t *node, int key);
+mlg_error_t mlg_hashtable_insert(mlg_hash_table_t *hash_table, mlg_hash_node_t *node, size_t key);
 mlg_error_t mlg_hashtable_remove(mlg_hash_node_t *node);
 
 #endif // MLG_INTRUSIVE_HASHTABLE_H
